Fixed POST /api/v1/tokens and POST /api/v1/users/reset 1. При вызове метода POST /api/v1/tokens если type = change_phone или reset_password, то проверяется, если пользователь с таким телефоном существует и он активирован. Если пользователь не существует, получишь ошибку {   "error": {     "code": "GEN-FORBIDDEN",     "http_code": 403,     "message": "user_not_found"   } } Если существует пользователь, но он не активирован получишь ошибку {   "error": {     "code": "GEN-FORBIDDEN",     "http_code": 403,     "message": "user_not_activated"   } }
2. Метод POST /api/v1/users/reset
Fixed - реально меняет пароль в БД
Added - возвращает токен авторизации (так же как и метод регистрации)
Пример:
{
  "data": {
    "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOjMzMSwiaXNzIjoiaHR0cDpcL1wvdmlsbGFnZS5kZXZcL2FwaVwvdjFcL3VzZXJzXC9yZXNldCIsImlhdCI6IjE0NDQ5OTM4MjgiLCJleHAiOiIxNDQ1MDgwMjI4IiwibmJmIjoiMTQ0NDk5MzgyOCIsImp0aSI6Ijk1NGQ3YWIzYmRiYmQwNDgwNDI3YjllMjY3OTI4ZGQ1In0.aEJAFRkcjWPalTsG1xyQ7H2m8bHl4KW3vmV3j9tstTk"
  }
}

+++ Modules/Village/Http/Controllers/Api/V1/UserController.php
@@ -134,16 +134,26 @@ class UserController extends ApiController
         unset($data['session'], $data['code']);
 
         $user = User::where(['phone' => $token['phone']])->first();
+        /** @var $user User */
         if (!$user) {
             return $this->response->errorNotFound('user_not_found');
         }
 
-        DB::transaction(function() use ($user, $token) {
-            $user->update(['phone' => $token['phone']]);
+        if (!$user->isActivated()) {
+            return $this->response->errorNotFound('user_not_activated');
+        }
+
+        $password = $data['password'];
+        $data['password'] = Hash::make($data['password']);
+
+        DB::transaction(function() use ($user, $token, $data) {
+            $user->update(['phone' => $token['phone'], 'password' => $data['password']]);
             $token->delete();
         });
 
-        return $this->response->withArray(['success' => true]);
+        $request::replace(['phone' => $token['phone'], 'password' => $password]);
+
+        return (new AuthController($this->response))->auth($request);
     }
 
 //    /**
